Fix tesla build when no tenant ID is provided
Contrary to the comment in kube-controllers/cmd/kube-controllers/cloud.go stating that the tenant ID is optional for cloud / tesla, it is currently required and the roles / users are modified to include the tenant ID. This means that currently the tesla build cannot be used for new calico cloud only features.

This commit makes the tenant ID truly optional for the tesla build, and creates the roles and es users as they would be in ee if the tenant ID is not provided.

+++ kube-controllers/pkg/controllers/authorization/cloud.go
@@ -33,7 +33,11 @@ var resourceNameToGlobalElasticsearchRoles = map[string]string{
 }
 
 func formatRoleName(name string) string {
-	return fmt.Sprintf("%s_%s", name, tenantID)
+	if tenantID != "" {
+		return fmt.Sprintf("%s_%s", name, tenantID)
+	} else {
+		return name
+	}
 }
 
 // resync removes all elasticsearch native users with prefix `tigera-k8s` that doesn't have an entry in user cache
@@ -41,41 +45,49 @@ func formatRoleName(name string) string {
 // This is the Cloud/Tesla variant of this function which ignores any users that do not have the tenantID suffix
 // in their role names to avoid overwriting another tenants oidc users.
 func (n *nativeUserSynchronizer) resync() error {
-	users, err := n.esCLI.GetUsers()
-	if err != nil {
-		return err
-	}
+	if tenantID != "" {
+		users, err := n.esCLI.GetUsers()
+		if err != nil {
+			return err
+		}
 
-	for _, user := range users {
-		if strings.HasPrefix(user.Username, nativeUserPrefix) {
-			rolesNames := user.RoleNames()
-			// Skip deleting this user if it does not contain roles specific to this tenant.
-			if !strings.HasSuffix(rolesNames[0], tenantID) {
-				continue
-			}
-			subjectID := strings.TrimPrefix(user.Username, fmt.Sprintf("%s-", nativeUserPrefix))
-			if !n.userCache.Exists(subjectID) {
-				if err := n.esCLI.DeleteUser(elasticsearch.User{Username: user.Username}); err != nil {
-					return err
+		for _, user := range users {
+			if strings.HasPrefix(user.Username, nativeUserPrefix) {
+				rolesNames := user.RoleNames()
+				// Skip deleting this user if it does not contain roles specific to this tenant.
+				if !strings.HasSuffix(rolesNames[0], tenantID) {
+					continue
+				}
+				subjectID := strings.TrimPrefix(user.Username, fmt.Sprintf("%s-", nativeUserPrefix))
+				if !n.userCache.Exists(subjectID) {
+					if err := n.esCLI.DeleteUser(elasticsearch.User{Username: user.Username}); err != nil {
+						return err
+					}
 				}
 			}
 		}
+
+		subjects := n.userCache.SubjectIDs()
+		return n.synchronizeOIDCUsers(subjects)
 	}
 
-	subjects := n.userCache.SubjectIDs()
-	return n.synchronizeOIDCUsers(subjects)
+	return n.eeResync()
 }
 
 func (n *nativeUserSynchronizer) deleteEsUsers(esUsers map[string]elasticsearch.User) error {
-	for _, esUser := range esUsers {
-		rolesNames := esUser.RoleNames()
-		// Skip deleting this user if it does not contain roles specific to this tenant.
-		if len(rolesNames) != 0 && !strings.HasSuffix(rolesNames[0], tenantID) {
-			continue
+	if tenantID != "" {
+		for _, esUser := range esUsers {
+			rolesNames := esUser.RoleNames()
+			// Skip deleting this user if it does not contain roles specific to this tenant.
+			if len(rolesNames) != 0 && !strings.HasSuffix(rolesNames[0], tenantID) {
+				continue
+			}
+			if err := n.esCLI.DeleteUser(esUser); err != nil {
+				return err
+			}
 		}
-		if err := n.esCLI.DeleteUser(esUser); err != nil {
-			return err
-		}
+		return nil
 	}
-	return nil
+
+	return n.eeDeleteEsUsers(esUsers)
 }
